ComponentRegistry::getVariable no longer converts all exceptions thrown by ParsedExpressionVariable in to UnknownVariableException, since doing so can mask the true cause of parse errors; ParsedExpressionVariable now throws UnknownVariableException when appropriate.  ParsedExpressionVariable now trims whitespace from expressions before parsing, because parsing can fail if it doesn't.  Made it easier to attach a debugger to a running instance of MWorksCoreTestRunner.  Fixed a broken unit test.

+++ core/Core/Variables/tests/ParsedExpressionVariableTest.cpp
@@ -364,7 +364,7 @@ void ParsedExpressionVariableTestFixture::testRangeExpression() {
     CPPUNIT_ASSERT_EQUAL(8, int(values.size()));
     for (valueList::size_type i = 0; i < values.size(); i++) {
         CPPUNIT_ASSERT( values[i].isIntegerType() );
-        CPPUNIT_ASSERT_EQUAL((long long)(5-i), values[i].getLong());
+        CPPUNIT_ASSERT_EQUAL((long long)(5) - (long long)(i), values[i].getLong());
     }
     
     // start == stop
